+++ frontend/app/controllers/spree/orders_controller.rb
@@ -41,7 +41,7 @@ module Spree
     # Adds a new item to the order (creating a new order if none already exists)
     def populate
       populator = Spree::OrderPopulator.new(current_order(true), current_currency)
-      if populator.populate(params.slice(:products, :variants, :quantity))
+      if populator.populate(params[:variant_id], params[:quantity])
         current_order.ensure_updated_shipments
 
         respond_with(@order) do |format|